start proxyjs. remove old commands and junk. make pray a function
