--- commands/music/music.js
@@ -1,4 +1,0 @@
-
-export default function play(msg, splitMessage) {
-    msg.channel.send(`${msg.author}, your roll is : ${roll}/${splitMessage[1]}`);
-}
