--- functions/ping/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "ping",
-  "version": "1.0.0",
-  "description": "",
-  "main": "ping.js",
-  "scripts": {
-    "zip": "zip -r ../../exports/ping.zip *",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "tweetnacl": "^1.0.3"
-  }
-}
