+++ functions/pray/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "pray",
+  "version": "1.0.0",
+  "description": "",
+  "main": "pray.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
